% Generated by build script
\name{foia_agencies}
\alias{foia_agencies}
\title{Fetch FOIA agency components}
\usage{
foia_agencies()
}
\arguments{
\item{api_key}{API key for api.data.gov. DEMO_KEY works for testing.}
}
\value{tibble: id, agency_name, title, abbreviation, website}
\description{Fetch FOIA agency components}
